% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCreator.R
\name{constCharLength}
\alias{constCharLength}
\title{Constant length checker}
\usage{
constCharLength(vect)
}
\arguments{
\item{vect}{}
}
\value{
Boolean, True when all same length, False otherwise
}
\description{
function to check that the lengths of the characters in a vector are consistent
}
