% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCreator.R
\name{dateForm}
\alias{dateForm}
\title{Date form checker (PLACEHOLDER)}
\usage{
dateForm(vect)
}
\arguments{
\item{vect}{}
}
\value{
TRUE always due to placeholder status
}
\description{
Function to check that the object is a valid date format
}
