% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCreator.R
\name{guessDataType}
\alias{guessDataType}
\title{dataTypeGuesser}
\usage{
guessDataType(vect)
}
\arguments{
\item{vect}{}
}
\value{
One of an array of data types which can be in a dataset
}
\description{
Function to guess the datatypes of a given vector
}
