% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{irisUltra}
\alias{irisUltra}
\title{Example Ultra Data.Frame}
\format{A list of 4 objects, the origin anticipator, the metadata dictionary, the spoofed dataset and the data analysis stages}
\usage{
irisUltra
}
\description{
An Example ultra_df which can be used in testing and vignettes.
}
\keyword{datasets}
