% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCreator.R
\name{postForm}
\alias{postForm}
\title{Post form checker (PLACEHOLDER)}
\usage{
postForm(vect)
}
\arguments{
\item{vect}{}
}
\value{
TRUE always due to placeholder status
}
\description{
Function to check that the object is a valid postcode format
}
