% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spoofer.R
\name{ultraAggRandom}
\alias{ultraAggRandom}
\title{Constrained Random generator}
\usage{
ultraAggRandom(vect)
}
\arguments{
\item{vect}{the vector of values}
}
\value{

}
\description{
Constrained Random generator
}
