% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anticipators.R
\name{ultraAnticipatorDict}
\alias{ultraAnticipatorDict}
\title{Anticipator Dictionary function}
\usage{
ultraAnticipatorDict(type)
}
\arguments{
\item{type}{A data type supported by the dict. one of c("CSV", "RDS)}
}
\value{

}
\description{
Returns the anticipator generator function based on the data type given
}
