% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfMethods.R
\name{ultraCallStagedOutput}
\alias{ultraCallStagedOutput}
\title{Call Staged Output}
\usage{
ultraCallStagedOutput(stages)
}
\arguments{
\item{stages}{}
}
\description{
Calls the stages attribute of the ultra_df to generate a multi step origin call, allowing for extra steps to be used after the source
}
