% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spoofer.R
\name{ultraContinuousCheck}
\alias{ultraContinuousCheck}
\title{Continuous Column Checker}
\usage{
ultraContinuousCheck(ultra_df)
}
\arguments{
\item{ultra_df}{the Ultra df}
}
\description{
Continuous Column Checker
}
