% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anticipators.R
\name{ultraCSV}
\alias{ultraCSV}
\title{CSV anticipator generator}
\usage{
ultraCSV(path)
}
\arguments{
\item{path}{The path of the csv file}
}
\value{

}
\description{
Generates an anticipator for the .csv file at the given path
}
