% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anticipators.R
\name{ultraExcelRemote}
\alias{ultraExcelRemote}
\title{Remote Excel anticipator generator}
\usage{
ultraExcelRemote(path)
}
\arguments{
\item{path}{}
}
\description{
Generates an anticipator for the .rds file at the given path
}
