% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfMethods.R
\name{ultraGetMax}
\alias{ultraGetMax}
\title{Maximum value getter}
\usage{
ultraGetMax(ultra_df, column)
}
\arguments{
\item{ultra_df}{the ultra df}

\item{column}{the column to query}
}
\value{

}
\description{
Maximum value getter
}
