% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfMethods.R
\name{ultraGetMin}
\alias{ultraGetMin}
\title{Minimum value getter}
\usage{
ultraGetMin(ultra_df, column)
}
\arguments{
\item{ultra_df}{the ultra df}

\item{column}{the column to query}
}
\value{

}
\description{
Minimum value getter
}
