% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCreator.R
\name{ultraMetaCalc}
\alias{ultraMetaCalc}
\title{Columnwise Metadata query}
\usage{
ultraMetaCalc(dataset, colname_in_question)
}
\arguments{
\item{colname_in_question}{}
}
\description{
Columnwise Metadata query
}
