% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generators.R
\name{ultraMetaGenerator}
\alias{ultraMetaGenerator}
\title{Injecting a metadata structure into the ultra_df}
\usage{
ultraMetaGenerator(ultra_df, dict = NULL)
}
\arguments{
\item{dict}{}
}
\description{
Populates the data structure with the metadata of the underlying origin data
}
