% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generators.R
\name{ultraMetaInjector}
\alias{ultraMetaInjector}
\title{Injecting a metadata structure into the ultra_df}
\usage{
ultraMetaInjector(ultra_df)
}
\arguments{
\item{ultra_df}{}
}
\description{
Populates the data structure with the metadata of the underlying origin data
}
