% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spoofer.R
\name{ultraNameSpoof}
\alias{ultraNameSpoof}
\title{Random Name generator}
\usage{
ultraNameSpoof()
}
\description{
uses the embedded mtcars and state.name datasets to generate passable, if often amusing fake names in PII slots
}
