% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spoofer.R
\name{ultraNominalCheck}
\alias{ultraNominalCheck}
\title{Nominal Column Checker}
\usage{
ultraNominalCheck(ultra_df)
}
\arguments{
\item{ultra_df}{The ultra df}
}
\value{

}
\description{
Nominal Column Checker
}
