% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anticipators.R
\name{ultraRDS}
\alias{ultraRDS}
\title{RDS anticipator generator}
\usage{
ultraRDS(path)
}
\arguments{
\item{path}{The path of the RDS file}
}
\description{
Generates an anticipator for the .rds file at the given path
}
