% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spoofer.R
\name{ultraSingleSpoofRun}
\alias{ultraSingleSpoofRun}
\title{Single Spoof run on grouped data}
\usage{
ultraSingleSpoofRun(grouped_origin, continuous_cols)
}
\arguments{
\item{continuous_cols}{}
}
\description{
Single Spoof run on grouped data
}
