% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generators.R
\name{ultraSpoof}
\alias{ultraSpoof}
\title{Spoofer}
\usage{
ultraSpoof(ultra_df)
}
\arguments{
\item{ultra_df}{}
}
\description{
Function to inject the spoofed data into the data structure, cloning the metadata of the spoofed object
}
