% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spoofer.R
\name{ultraSpooferNew}
\alias{ultraSpooferNew}
\title{New spoofing functio}
\usage{
ultraSpooferNew(ultra_df)
}
\arguments{
\item{ultra_df}{the ultra df}
}
\value{

}
\description{
Function which uses the new devised method to spoof without introducing illegal combinations of values.
}
