% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCreator.R
\name{uniqueness}
\alias{uniqueness}
\title{Uniqueness function}
\usage{
uniqueness(vect)
}
\arguments{
\item{vect}{}
}
\value{
Value for the uniqueness between 1/n and 1, 1 being entirely unique and n being the number of elements in vect
}
\description{
A function to determine the uniqueness of a vector
}
