% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generators.R
\name{ultraMeta}
\alias{ultraMeta}
\title{Injecting a metadata structure into the ultra_df}
\usage{
ultraMeta(ultra_df, dict = NULL)
}
\arguments{
\item{ultra_df}{the ultra df object}

\item{dict}{The metadat dictionary}
}
\value{

}
\description{
Populates the data structure with the metadata of the underlying origin data
}
