% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCreator.R
\name{ultraMetaCorrector}
\alias{ultraMetaCorrector}
\title{Metadata corrector}
\usage{
ultraMetaCorrector(ultra_df, cols_to_correct)
}
\arguments{
\item{ultra_df}{The ultra df}

\item{cols_to_correct}{the columns to correct in the metadict}
}
\value{

}
\description{
Function to interactively update the contents of the metadata tibble by columnname
}
