% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCreator.R
\name{ultraMetaGenerator}
\alias{ultraMetaGenerator}
\title{Creating a metadata structure for the ultra_df}
\usage{
ultraMetaGenerator(dat)
}
\arguments{
\item{ultra_df}{The ultra df}
}
\value{

}
\description{
Generates the metadata table for future use and underlying the spoofing mechanism
}
