% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spoofer.R
\name{ultraRandoColumn}
\alias{ultraRandoColumn}
\title{Random column generator}
\usage{
ultraRandoColumn(ultra_df, Column_in_question)
}
\arguments{
\item{ultra_df}{The ultra df}

\item{Column_in_question}{The column for the function}
}
\value{

}
\description{
Random column generator
}
