% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spoofer.R
\name{ultraSingleSpoofRun}
\alias{ultraSingleSpoofRun}
\title{Single Spoof run on grouped data}
\usage{
ultraSingleSpoofRun(grouped_origin, continuous_cols)
}
\arguments{
\item{grouped_origin}{the origin data that has been grouped already}

\item{continuous_cols}{the continuous columns}
}
\value{

}
\description{
Single Spoof run on grouped data
}
