% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generators.R
\name{ultraWrapper}
\alias{ultraWrapper}
\title{Wrapper}
\usage{
ultraWrapper(origin_location, type)
}
\arguments{
\item{type}{The type of file found at the path}

\item{origin}{the path of the source file}
}
\description{
The main function to wrap up dataframes into the ultrawrapper format
}
